Third update to the project, ready to be used.

+++ src/facade/AdminFacade.java
@@ -8,6 +8,7 @@ import dataBaseDAO.CompanyDBDAO;
 import dataBaseDAO.CouponDBDAO;
 import dataBaseDAO.CustomerDBDAO;
 import exceptions.CouponSystemException;
+import exceptions.FacadeException;
 
 public class AdminFacade implements CouponClientFacade {
 	
@@ -16,10 +17,12 @@ public class AdminFacade implements CouponClientFacade {
 	private CouponDBDAO coupDBDAO;
 
 	/**
-	 * Constructor for the admin facade.
+	 * Constructor for the Admin facade.
+	 * 
+	 * @throws CouponSystemException if there were issues during the data access objects creation.
 	 * 
 	 */
-	public AdminFacade() {
+	public AdminFacade() throws CouponSystemException {
 		compDBDAO = new CompanyDBDAO();
 		custDBDAO = new CustomerDBDAO(); 
 		coupDBDAO = new CouponDBDAO(); 
@@ -30,13 +33,14 @@ public class AdminFacade implements CouponClientFacade {
 	 * Creates a company in the database.
 	 * 
 	 * @param comp is the company to be added.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if a company with such name already exists.
 	 */
 	public void createCompany(Company comp) throws CouponSystemException {
 		if(compDBDAO.getCompanyByName(comp.getCompName())==null){
 			compDBDAO.createCompany(comp);
 		}else {
-			throw new CouponSystemException("Error: Name already taken.");
+			throw new FacadeException("Error: Name already taken.");
 		}
 		
 		
@@ -46,13 +50,39 @@ public class AdminFacade implements CouponClientFacade {
 	 * Removes a company from the database.
 	 * 
 	 * @param comp is the company to be removed.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if no such company was found.
 	 */
 	public void removeCompany(Company comp) throws CouponSystemException {
-		coupDBDAO.removeCustomerCoupons(comp.getCoupons());
-		coupDBDAO.removeCoupons(comp.getCoupons());
-		coupDBDAO.removeCompanyCoupons(comp.getCoupons());
-		compDBDAO.removeCompany(comp);		
+		if(compDBDAO.getCompany(comp.getId()) != null) {
+			coupDBDAO.removeCustomerCoupons(comp.getCoupons());
+			coupDBDAO.removeCoupons(comp.getCoupons());
+			coupDBDAO.removeCompanyCoupons(comp.getCoupons());
+			compDBDAO.removeCompany(comp);			
+		} else {
+			throw new FacadeException("Error: no such company exists.");
+		}
+		
+	}
+	
+	/**
+	 * Removes a list of companies from the database.
+	 * 
+	 * @param companies is the list of companies to be removed.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if no such company was found.
+	 */
+	public void removeCompanies(ArrayList<Company> companies) throws CouponSystemException {
+		for (Company comp : companies) {
+			if(compDBDAO.getCompany(comp.getId()) != null) {
+				coupDBDAO.removeCustomerCoupons(comp.getCoupons());
+				coupDBDAO.removeCoupons(comp.getCoupons());
+				coupDBDAO.removeCompanyCoupons(comp.getCoupons());
+				compDBDAO.removeCompany(comp);			
+			} else {
+				throw new FacadeException("Error: A company from the list doesn't exists.");
+			}	
+		}
 		
 	}
 	
@@ -60,14 +90,15 @@ public class AdminFacade implements CouponClientFacade {
 	 * Updates a company in the database.
 	 * 
 	 * @param comp is the new version of the company that is to be updated.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if the name or id were incorrect.
 	 */
 	public void updateCompany(Company comp) throws CouponSystemException {
 		if (compDBDAO.getCompanyByName(comp.getCompName()) != null
 				&& compDBDAO.getCompanyByName(comp.getCompName()).getId() == comp.getId()) {
 			compDBDAO.updateCompany(comp);
 		} else {
-			throw new CouponSystemException("Error: Cannot update Company. Incorrect name.");
+			throw new FacadeException("Error: Cannot update Company. Incorrect name.");
 		}
 
 	}
@@ -97,7 +128,8 @@ public class AdminFacade implements CouponClientFacade {
 	 * Creates a customer in the database.
 	 * 
 	 * @param cust is the customer to be added.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if customer with such name already exists.
 	 */
 	public void createCustomer(Customer cust) throws CouponSystemException {
 		if(custDBDAO.getCustomerByName(cust.getCustName())==null){
@@ -113,11 +145,35 @@ public class AdminFacade implements CouponClientFacade {
 	 * Removes a customer from the database.
 	 * 
 	 * @param cust is the customer to be removed.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if no such customer was found.
 	 */
 	public void removeCustomer(Customer cust) throws CouponSystemException {
-		coupDBDAO.removeCustomerCoupons(cust.getCoupons());
-		custDBDAO.removeCustomer(cust);
+		if(custDBDAO.getCustomer(cust.getId()) != null) {
+			coupDBDAO.removeCustomerCoupons(cust.getCoupons());
+			custDBDAO.removeCustomer(cust);
+		} else {
+			throw new FacadeException("Error: no such customer exists.");
+		}
+		
+		
+	}
+	/**
+	 * Removes a list customers from the database.
+	 * 
+	 * @param customers is the customer list to be removed.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if no such customer was found.
+	 */
+	public void removeCustomers(ArrayList<Customer> customers) throws CouponSystemException {
+		for (Customer cust : customers) {
+			if(custDBDAO.getCustomer(cust.getId()) != null) {
+				coupDBDAO.removeCustomerCoupons(cust.getCoupons());
+				custDBDAO.removeCustomer(cust);
+			} else {
+				throw new FacadeException("Error: A customer from the list doesn't exists.");
+			}	
+		}
 		
 		
 	}
@@ -126,14 +182,15 @@ public class AdminFacade implements CouponClientFacade {
 	 * Updates a customer in the database.
 	 * 
 	 * @param cust is the new version of the customer that is to be updated.
-	 * @throws CouponSystemException if there were issues during method runtime.
+	 * @throws CouponSystemException if there were issues during method runtime,
+	 * or if the name or id were incorrect.
 	 */
 	public void updateCustomer(Customer cust) throws CouponSystemException {
 		if (custDBDAO.getCustomerByName(cust.getCustName()) != null
 				&& custDBDAO.getCustomerByName(cust.getCustName()).getId() == cust.getId()) {
 			custDBDAO.updateCustomer(cust);
 		} else {
-			throw new CouponSystemException("Error: Cannot update Customer. Incorrect name.");
+			throw new FacadeException("Error: Cannot update Customer. Incorrect name.");
 		}
 
 	}
